requirements.txt
# Core Data & Analytics
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
pandas-ta==0.3.14b0
scikit-learn==1.3.2
scipy==1.11.4

# Machine Learning & Deep Learning
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
stable-baselines3==2.2.1
gymnasium==0.29.1

# Reinforcement Learning & Optimization
optuna==3.5.0
ray[tune]==2.8.1

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.1

# Database & Caching
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# API & Data Sources
requests==2.31.0
websocket-client==1.7.0
finnhub-python==2.4.20
alpaca-trade-api==3.1.1

# Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1

# Backtesting
backtrader==1.9.78.123

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.17.0
dash==2.16.1

# Notifications
python-telegram-bot==20.7
smtplib-ssl==1.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
schedule==1.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Docker & Deployment
gunicorn==21.2.0
uvicorn==0.25.0
fastapi==0.108.0